day35/autoquant_day35/requirements.txt
alpaca-py>=0.13.0
rich>=13.7.0
python-dotenv>=1.0.0
